requirements.txt
gradescope-utils>=0.2.6
pyyaml